# hackathon-II/hackathon-todo-phase3/specs/1-fullstack-todo-app/contracts/tasks-openapi.yaml
openapi: 3.0.3
info:
  title: Todo Application API
  description: API for managing user tasks in the Todo application
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/v1
    description: Production server

paths:
  /auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  minLength: 8
                  description: User's password
                name:
                  type: string
                  maxLength: 100
                  description: User's name
              required:
                - email
                - password
                - name
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT token for authentication
        '400':
          description: Invalid input
        '409':
          description: Email already exists

  /auth/signin:
    post:
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  description: User's password
              required:
                - email
                - password
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT token for authentication
        '400':
          description: Invalid input
        '401':
          description: Invalid credentials

  /{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user's ID
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending, completed]
          description: Filter tasks by status
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - cannot access other users' tasks
      security:
        - BearerAuth: []

    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - cannot access other users' tasks
      security:
        - BearerAuth: []

  /{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task's ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - cannot access other users' tasks
        '404':
          description: Task not found
      security:
        - BearerAuth: []

    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - cannot access other users' tasks
        '404':
          description: Task not found
      security:
        - BearerAuth: []

    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task's ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                  type: string
                  example: "Task deleted"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - cannot access other users' tasks
        '404':
          description: Task not found
      security:
        - BearerAuth: []

  /{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The task's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: New completion status
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  completed:
                    type: boolean
                  updated_at:
                    type: string
                    format: date-time
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - cannot access other users' tasks
        '404':
          description: Task not found
      security:
        - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 100
      required:
        - id
        - email
        - name

    Task:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 1000
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at

    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 1000
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 1000